blog: saving html file

+++ blog/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 from django import forms
 from django.forms import TextInput, Textarea
+from django.core.files.base import ContentFile
 from .models import Post, PostImage
 # Register your models here.
 
@@ -17,7 +18,8 @@ class PostAdminForm(forms.ModelForm):
             'body': Textarea(attrs={'rows': 30, 'cols': 100}),
             'title': TextInput(attrs={'size': 40}),
         }
-        exclude = ('html_file', 'created_time', 'modified_time')
+        exclude = ('toc', 'excerpect', 'created_time',
+                   'modified_time', 'count_read')
 
 
 class PostAdmin(admin.ModelAdmin):
@@ -28,11 +30,18 @@ class PostAdmin(admin.ModelAdmin):
     search_fields = ['title', 'author']
 
     def save_model(self, request, obj, form, change):
-        # TODO: update same name images
         if obj:
-            if any(map(lambda x: x in form.changed_data,
-                       ('title', 'body'))):
+            if change and any(map(lambda x: x in form.changed_data,
+                                  ('title', 'body', 'html_file'))):
                 obj.html_file.delete()
+
+            if 'html_file' in form.changed_data:
+                html_file = request.FILES['html_file']
+                with html_file.open() as f:
+                    obj.html_file.save(obj.title+'.html', ContentFile(
+                        f.read()), save=False)
+                    obj.html_file.close()
+
             obj.save()
